+++ src/test/kotlin/net/pwall/json/schema/codegen/test/kotlin/TestAdditionalPropertiesSchemaExtra.kt
@@ -32,20 +32,7 @@ class TestAdditionalPropertiesSchemaExtra(
 
     val extra: String by cg_map
 
-    override fun toString() = buildString {
-        append("TestAdditionalPropertiesSchemaExtra(")
-        if (cg_map.isNotEmpty()) {
-            var count = 0
-            cg_map.entries.forEach { (key, value) ->
-                append(key)
-                append('=')
-                append(value)
-                if (++count < cg_map.size)
-                    append(", ")
-            }
-        }
-        append(')')
-    }
+    override fun toString(): String = "TestAdditionalPropertiesSchemaExtra(${cg_map.entries.joinToString { "${it.key}=${it.value}" }})"
 
     override fun equals(other: Any?): Boolean = this === other || other is TestAdditionalPropertiesSchemaExtra && cg_map == other.cg_map