+++ src/test/kotlin/net/pwall/json/schema/codegen/test/kotlin/TestAdditionalPropertiesSchemaExtra2.kt
@@ -41,20 +41,7 @@ class TestAdditionalPropertiesSchemaExtra2(
     /** Extra field 2 */
     val extra2: Int by cg_map
 
-    override fun toString() = buildString {
-        append("TestAdditionalPropertiesSchemaExtra2(")
-        if (cg_map.isNotEmpty()) {
-            var count = 0
-            cg_map.entries.forEach { (key, value) ->
-                append(key)
-                append('=')
-                append(value)
-                if (++count < cg_map.size)
-                    append(", ")
-            }
-        }
-        append(')')
-    }
+    override fun toString(): String = "TestAdditionalPropertiesSchemaExtra2(${cg_map.entries.joinToString { "${it.key}=${it.value}" }})"
 
     override fun equals(other: Any?): Boolean = this === other || other is TestAdditionalPropertiesSchemaExtra2 && cg_map == other.cg_map