+++ src/test/kotlin/net/pwall/json/schema/codegen/test/kotlin/TestAdditionalPropertiesTrue.kt
@@ -16,20 +16,7 @@ class TestAdditionalPropertiesTrue(
     private val cg_map: Map<String, Any?>
 ) : Map<String, Any?> by cg_map {
 
-    override fun toString() = buildString {
-        append("TestAdditionalPropertiesTrue(")
-        if (cg_map.isNotEmpty()) {
-            var count = 0
-            cg_map.entries.forEach { (key, value) ->
-                append(key)
-                append('=')
-                append(value)
-                if (++count < cg_map.size)
-                    append(", ")
-            }
-        }
-        append(')')
-    }
+    override fun toString(): String = "TestAdditionalPropertiesTrue(${cg_map.entries.joinToString { "${it.key}=${it.value}" }})"
 
     override fun equals(other: Any?): Boolean = this === other || other is TestAdditionalPropertiesTrue && cg_map == other.cg_map