+++ src/test/kotlin/net/pwall/json/schema/codegen/test/kotlin/TestApNestedTrue.kt
@@ -20,20 +20,7 @@ data class TestApNestedTrue(
         private val cg_map: Map<String, Any?>
     ) : Map<String, Any?> by cg_map {
 
-        override fun toString() = buildString {
-            append("Nested(")
-            if (cg_map.isNotEmpty()) {
-                var count = 0
-                cg_map.entries.forEach { (key, value) ->
-                    append(key)
-                    append('=')
-                    append(value)
-                    if (++count < cg_map.size)
-                        append(", ")
-                }
-            }
-            append(')')
-        }
+        override fun toString(): String = "Nested(${cg_map.entries.joinToString { "${it.key}=${it.value}" }})"
 
         override fun equals(other: Any?): Boolean = this === other || other is Nested && cg_map == other.cg_map