+++ src/test/kotlin/net/pwall/json/schema/codegen/test/kotlin/TestApTrueExtra.kt
@@ -26,20 +26,7 @@ class TestApTrueExtra(
 
     val extra: String by cg_map
 
-    override fun toString() = buildString {
-        append("TestApTrueExtra(")
-        if (cg_map.isNotEmpty()) {
-            var count = 0
-            cg_map.entries.forEach { (key, value) ->
-                append(key)
-                append('=')
-                append(value)
-                if (++count < cg_map.size)
-                    append(", ")
-            }
-        }
-        append(')')
-    }
+    override fun toString(): String = "TestApTrueExtra(${cg_map.entries.joinToString { "${it.key}=${it.value}" }})"
 
     override fun equals(other: Any?): Boolean = this === other || other is TestApTrueExtra && cg_map == other.cg_map