+++ src/test/kotlin/net/pwall/json/schema/codegen/test/kotlin/TestApTrueExtraPattern.kt
@@ -30,20 +30,7 @@ class TestApTrueExtraPattern(
 
     val extra: String by cg_map
 
-    override fun toString() = buildString {
-        append("TestApTrueExtraPattern(")
-        if (cg_map.isNotEmpty()) {
-            var count = 0
-            cg_map.entries.forEach { (key, value) ->
-                append(key)
-                append('=')
-                append(value)
-                if (++count < cg_map.size)
-                    append(", ")
-            }
-        }
-        append(')')
-    }
+    override fun toString(): String = "TestApTrueExtraPattern(${cg_map.entries.joinToString { "${it.key}=${it.value}" }})"
 
     override fun equals(other: Any?): Boolean = this === other || other is TestApTrueExtraPattern && cg_map == other.cg_map