+++ src/test/kotlin/net/pwall/json/schema/codegen/test/kotlin/TestApTrueMinMax.kt
@@ -20,20 +20,7 @@ class TestApTrueMinMax(
         require(cg_map.size in 2..5) { "Number of properties not in range 2..5 - ${cg_map.size}" }
     }
 
-    override fun toString() = buildString {
-        append("TestApTrueMinMax(")
-        if (cg_map.isNotEmpty()) {
-            var count = 0
-            cg_map.entries.forEach { (key, value) ->
-                append(key)
-                append('=')
-                append(value)
-                if (++count < cg_map.size)
-                    append(", ")
-            }
-        }
-        append(')')
-    }
+    override fun toString(): String = "TestApTrueMinMax(${cg_map.entries.joinToString { "${it.key}=${it.value}" }})"
 
     override fun equals(other: Any?): Boolean = this === other || other is TestApTrueMinMax && cg_map == other.cg_map