+++ src/test/kotlin/net/pwall/json/schema/codegen/test/kotlin/TestApTruePatternValid.kt
@@ -25,20 +25,7 @@ class TestApTruePatternValid(
         }
     }
 
-    override fun toString() = buildString {
-        append("TestApTruePatternValid(")
-        if (cg_map.isNotEmpty()) {
-            var count = 0
-            cg_map.entries.forEach { (key, value) ->
-                append(key)
-                append('=')
-                append(value)
-                if (++count < cg_map.size)
-                    append(", ")
-            }
-        }
-        append(')')
-    }
+    override fun toString(): String = "TestApTruePatternValid(${cg_map.entries.joinToString { "${it.key}=${it.value}" }})"
 
     override fun equals(other: Any?): Boolean = this === other || other is TestApTruePatternValid && cg_map == other.cg_map